Prueba de configuracion de pines desde odoo

+++ models/led.py
@@ -3,7 +3,7 @@ from openerp import models, fields, api, exceptions
 # from datetime import date
 #import wiringpi2
 # from subprocess import check_output,call,CalledProcessError
-# import os
+import subprocess
 
 
 
@@ -21,39 +21,66 @@ class cubie_odoo_led(models.Model):
         readonly=True,
     )
 
+    def gpio_enable(self, pin):
+        gpio_existe = False
+        try:
+            archivos = subprocess.Popen(['ls','/sys/class/gpio/'], stdout=subprocess.PIPE)
+            ls_output = archivos.stdout.read()
+            for file in ls_output.split('\n'):
+                if file == 'gpio17_pg9':
+                    gpio_existe = True
+            if gpio_existe == False:
+                try:
+                    #Crear el enlace de configuracion del pin
+                    export = open('/sys/class/gpio/export','w')
+                    export.write(pin)
+                    cerrado = False
+                    while(cerrado == False):
+                        try:
+                            export.close()
+                            cerrado = True
+                        except IOError:
+                            print "Aun no se puede cerrar el archivo {}".format(export)
+                except Exception as e:
+                    print 'Error {}'.format(e)
+                    message = 'Hubo un problema habilitando el pin.'
+                    raise exceptions.Warning(message, str(e))
+
+            #Se configura como salida el pin
+            try:
+                path = '/sys/class/gpio/gpio' + pin + '_pg9/direction'
+                direction = open (path,'w')
+                direction.write('out')
+                cerrado = False
+                while(cerrado == False):
+                    try:
+                        direction.close()
+                        cerrado = True
+                    except IOError:
+                        print "Aun no se puede cerrar el archivo {}".format(direction)
+                try:
+                    subprocess.call('sudo chown -R odoo:odoo /sys/class/gpio/gpio17_pg9',shell = True)
+                except Exception as e:
+                    print 'Error {}'.format(e)
+                    message = 'Hubo un problema cambiando propietario a los archivos de configuracion del pin.'
+                    raise exceptions.Warning(message, str(e))
+            except Exception as e:
+                print 'Error {}'.format(e)
+                message = 'Hubo un problema configurando el pin.'
+                raise exceptions.Warning(message, str(e))
+        except Exception as e:
+            print 'Error {}'.format(e)
+            message = 'Hubo un problema prendiendo el LED.'
+            raise exceptions.Warning(message, str(e))
+
 
     @api.model
     def create(self, values):
-#         Crear el enlace de configuracion del pin
         try:
-            export= open('/sys/class/gpio/export','w')
-            export.write(self.led_pin)
-            cerrado = False
-            while(cerrado == False):
-                try:
-                    export.close()
-                    cerrado = True
-                except IOError:
-                    print "Aun no se puede cerrar el archivo {}".format(export)
-        except IOError:
-            message = 'Hubo un problema habilitando el pin.'
-            print(message)
-
-#         Configurarlo como salida
-        try:
-            path = '/sys/class/gpio/gpio' + self.led_pin + '_pg9/direction'
-            direction = open (path,'w')
-            direction.write('out')
-            cerrado = False
-            while(cerrado == False):
-                try:
-                    direction.close()
-                    cerrado = True
-                except IOError:
-                    print "Aun no se puede cerrar el archivo {}".format(direction)
+            self.gpio_enable(self.led_pin)
         except Exception as e:
             print 'Error {}'.format(e)
-            message = 'Hubo un problema configurando el pin.'
+            message = 'Hubo un problema Configurando el pin.'
             raise exceptions.Warning(message, str(e))
         return super(cubie_odoo_led, self).create(values)
 
@@ -62,6 +89,19 @@ class cubie_odoo_led(models.Model):
     @api.one
     def action_prender(self):
         self.state = 'prendido'
+        try:
+            gpio_existe = False
+            archivos = subprocess.Popen(['ls','/sys/class/gpio/'], stdout=subprocess.PIPE)
+            ls_output = archivos.stdout.read()
+            for file in ls_output.split('\n'):
+                if file == 'gpio17_pg9':
+                    gpio_existe = True
+            if gpio_existe == False:
+                self.gpio_enable(self.led_pin)
+        except Exception as e:
+            print 'Error {}'.format(e)
+            message = 'Hubo un problema configurando el pin.'
+            raise exceptions.Warning(message, str(e))
         try:
             path = '/sys/class/gpio/gpio' + self.led_pin + '_pg9/value'
             value= open (path,'w')
@@ -73,8 +113,6 @@ class cubie_odoo_led(models.Model):
                     cerrado = True
                 except IOError:
                     print "Aun no se puede cerrar el archivo {}".format(value)
-            
-
         except Exception as e:
             print 'Error {}'.format(e)
             message = 'Hubo un problema prendiendo el LED.'
@@ -84,6 +122,19 @@ class cubie_odoo_led(models.Model):
     @api.one
     def action_apagar(self):
         self.state = 'apagado'
+        try:
+            gpio_existe = False
+            archivos = subprocess.Popen(['ls','/sys/class/gpio/'], stdout=subprocess.PIPE)
+            ls_output = archivos.stdout.read()
+            for file in ls_output.split('\n'):
+                if file == 'gpio17_pg9':
+                    gpio_existe = True
+            if gpio_existe == False:
+                self.gpio_enable(self.led_pin)
+        except Exception as e:
+            print 'Error {}'.format(e)
+            message = 'Hubo un problema configurando el pin.'
+            raise exceptions.Warning(message, str(e))
         try:
             path = '/sys/class/gpio/gpio' + self.led_pin + '_pg9/value'
             value= open (path,'w')
